trading_bot/requirements.txt
# Trading Bot Dependencies

# Exchange API
ccxt>=4.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment
python-dotenv>=1.0.0

# API server (for dashboard)
fastapi>=0.100.0
uvicorn>=0.23.0

# Azure OpenAI (for LLM analysis)
openai>=1.0.0

# Notifications
python-telegram-bot>=20.0

# Scheduling
schedule>=1.2.0

# Testing
pytest>=7.0.0
